Refactor Projectlinker into separate components
